requirements.txt
keyboard==0.13.5
packaging==25.0
psutil==7.0.0
pymem==1.14.0
pynput==1.8.1
PyQt6==6.9.0
PyQt6_sip==13.10.0
python_dateutil==2.9.0.post0
pywin32==310
Requests==2.32.3
watchdog==6.0.0
PyGetWindow==0.0.9
orjson==3.10.18